+++ activerecord/test/models/book.rb
@@ -2,8 +2,10 @@ class Book < ActiveRecord::Base
   has_many :authors
 
   has_many :citations, :foreign_key => 'book1_id'
-  has_many :references, -> { distinct }, :through => :citations, :source => :reference_of
+  has_many :references, -> { distinct }, through: :citations, source: :reference_of
 
   has_many :subscriptions
-  has_many :subscribers, :through => :subscriptions
+  has_many :subscribers, through: :subscriptions
+  
+  enum status: %i( proposed written published )
 end