+++ activerecord/test/schema/schema.rb
@@ -94,6 +94,7 @@ ActiveRecord::Schema.define do
   create_table :books, :force => true do |t|
     t.integer :author_id
     t.column :name, :string
+    t.column :status, :integer, default: 0
   end
 
   create_table :booleans, :force => true do |t|